pyproject.toml
[project]
name = "pyrfu-optimization"
readme = "README.rst"
authors = [{ name = "Louis RICHARD", email = "loui.richard@irf.se" }]
license = { file = "LICENSE.txt" }
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Other Environment",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft",
    "Operating System :: Microsoft :: MS-DOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "pyrfu",
]
requires-python = ">=3.9,<3.13"

[project.optional-dependencies]
git = [
    "pre-commit",
]
style = [
    "black",
    "flake8",
    "isort",
    "pylint",
]
tests = [
    "ddt",
    "pytest",
    "pytest-cov",
]
type-checking = [
    "mypy",
]
docs = [
    "nbsphinx>=0.9.2",
    "numpydoc>=1.5.0",
    "pydata-sphinx-theme>=0.13.0",
    "sphinx>=7.0.0",
    "sphinx-codeautolink>=0.15.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-gallery>=0.13.0",
    "sphinxcontrib-apidoc>=0.3.0",
]
dev = [
    "pyrfu[docs]",
    "pyrfu[git]",
    "pyrfu[style]",
    "pyrfu[tests]",
    "pyrfu[type-checking]",
]

[project.urls]
homepage = "https://github.com/louis-richard/pyrfu-optimization"
source = "https://github.com/louis-richard/pyrfu-optimization"
issue-tracker = "https://github.com/louis-richard/pyrfu-optimization/issues"


[tool.pytest.ini_options]
log_file = "pyrfu-optimization-tests.log"
log_file_mode = "w"
log_file_format = "%(asctime)s %(levelname)s %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = [
    "ignore::RuntimeWarning",
]

[tool.black]
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist

  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
)/'''

[tool.isort]
profile = "black"

[tool.pylint."MESSAGES CONTROL"]
disable = """
    missing-module-docstring,
    too-many-function-args,
    too-many-positional-arguments,
    too-many-arguments,
    too-many-locals,
    too-many-lines,
    too-many-statements,
    too-many-branches,
    too-many-nested-blocks,
    invalid-name,
    duplicate-code,
    not-an-iterable,
    fixme
"""
ignore = """
    .git,
    .mypy_cache,
    .venv,
"""

[tool.mypy]
enable_error_code = "redundant-self"
files = "scripts"
show_error_codes = true
show_error_context = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

# Much of the numerical computing stack doesn't have type annotations yet.
[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "pyrfu.*",
]

# This then excludes some modules from the above list. (So ideally we remove
# from here in time...)
[[tool.mypy.overrides]]
check_untyped_defs = false
# module = [
# ]
